% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\docType{package}
\name{rmd-package}
\alias{rmd}
\alias{rmd-package}
\title{rmd: Easily Install and Load the R Markdown Family}
\description{
The 'rmd' package manages multiple R markdown packages. These R markdown packages include currently 'rmarkdown', 'knitr', 'bookdown', 'bookdownplus', 'blogdown', 'rticles',  'tinytex', 'xaringan', 'citr',  and 'mindr'. They can be installed and loaded in a single step with the 'rmd' package. The conflicts between these packages are evaluated as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pzhaonet/rmd}
  \item Report bugs at \url{https://github.com/pzhaonet/rmd/issues}
}

}
\author{
\strong{Maintainer}: Peng Zhao \email{pzhao@pzhao.net}

}
\keyword{internal}
