% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{rmd_conflicts}
\alias{rmd_conflicts}
\title{Conflicts between the rmd and other packages}
\usage{
rmd_conflicts()
}
\description{
This function lists all the conflicts between packages in the rmd
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
rmd_conflicts()
}
