% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{rmd_deps}
\alias{rmd_deps}
\title{List all rmd dependencies}
\usage{
rmd_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
rmd packages.}
}
\description{
List all rmd dependencies
}
