% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmd_packages}
\alias{rmd_packages}
\title{List all packages in the rmd}
\usage{
rmd_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include rmd in the list?}
}
\description{
List all packages in the rmd
}
\examples{
rmd_packages()
}
