% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{rmd_update}
\alias{rmd_update}
\title{Update rmd packages}
\usage{
rmd_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
rmd packages.}
}
\description{
This will check to see if all rmd packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
rmd_update()
}
}
