% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{rmdAddin}
\alias{rmdAddin}
\title{Addin for browsing and installing RStudio addins}
\usage{
rmdAddin()
}
\description{
This addin allows you to interactively browse through the list of existing
addins, see which ones you already have installed, and let you
install/uninstall the corresponding package of each addin. This addin can be
invoked from RStudio's "Addins" menu.
}
\examples{
if (interactive()) {
    rmdAddin()
}
}
